Setup ng-zorro in dark theme
