+++ client/src/app/app.html
@@ -1,0 +1,8 @@
+<!--
+* @Author                : Jbristhuille<jbristhuille@gmail.com>
+* @CreatedDate           : 2025-06-03 17:24:00
+* @LastEditors           : Jbristhuille<jbristhuille@gmail.com>
+* @LastEditDate          : 2025-06-03 17:41:45
+-->
+
+<router-outlet></router-outlet>
