+++ client/src/app/app.scss
@@ -1,6 +1,7 @@
 /*****************************************************************************
  * @Author                : Jbristhuille<jbristhuille@gmail.com>             *
- * @CreatedDate           : 2025-06-03 17:37:50                              *
+ * @CreatedDate           : 2025-06-03 17:24:49                              *
  * @LastEditors           : Jbristhuille<jbristhuille@gmail.com>             *
- * @LastEditDate          : 2025-06-03 17:40:36                              *
+ * @LastEditDate          : 2025-06-03 17:25:05                              *
  ****************************************************************************/
+
